third_party/WebKit/LayoutTests/editing/selection/leak-document-with-selection-inside-expected.txt

This test asserts that document doesn't leak when a selection is created inside the document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS The difference of counter "numberOfLiveDocuments" before and after the cycle is under the threshold of 0.
PASS successfullyParsed is true

TEST COMPLETE
